gracefully exit when table has more than 64K stuff
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/2809



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@11883 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/tools/genrb/reslist.c
@@ -916,7 +916,7 @@ uint16_t bundle_addtag(struct SRBRoot *bundle, const char *tag, UErrorCode *stat
         return (uint16_t) - 1;
     }
 
-    keypos = bundle->fKeyPoint;
+    keypos = (uint16_t)bundle->fKeyPoint;
 
     bundle->fKeyPoint += (uint16_t) (uprv_strlen(tag) + 1);
